daml2ts: Move types for create/archive events into daml-json-types
Currently, they are in `daml-ledger-fetch`, which doesn't make sense since
they are not bound to ledger clients implemented using fetch.
